pyproject.toml
[project]
name = "artisan-engine"
version = "0.1.0"
description = "A production-grade, OpenAI-compatible API layer for local LLMs with guaranteed structured output."
license = "Apache-2.0"
readme = "README.md"
requires-python = ">=3.12"
homepage = "https://github.com/aafre/artisan-engine"
repository = "https://github.com/aafre/artisan-engine"

classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "llama-cpp-python>=0.3.14",
    "numba>=0.61.2",
    "outlines>=1.2.1",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "python-multipart>=0.0.12",
    "httpx>=0.28.1",
    "python-dotenv>=1.0.0",
    "openai>=1.99.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.24.0",
    "ruff>=0.7.0",
]

[tool.ruff]
# Ruff configuration
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "-ra",
]
